Digital-Currency-Analysis-System/requirements.txt
certifi==2018.10.15
chardet==3.0.4
idna==2.7
kafka-python==1.4.3
requests==2.20.0
schedule==0.5.0
urllib3==1.24.1
